cli is finished as of now

+++ src/app/cli/cli.component.ts
@@ -2,7 +2,10 @@ import { Component, OnInit, AfterViewInit, ViewChild, ElementRef, ViewEncapsulat
 import { NgTerminal} from 'ng-terminal';
 import { Terminal} from 'xterm';
 import { Router } from '@angular/router';
-import { ClassStorageService } from '../class-storage.service';
+import { ClassStorageService, fullClass } from '../class-storage.service';
+import { ArrayType, collectExternalReferences } from '@angular/compiler';
+import { getMatTooltipInvalidPositionError } from '@angular/material';
+import { delay } from 'q';
 
 @Component({
   selector: 'app-cli',
@@ -21,16 +24,16 @@ export class CliComponent implements OnInit, AfterViewInit {
   @HostListener('document:keyup', ['$event'])
   handleDeleteKeyboardEvent(event: KeyboardEvent) {
     if(event.keyCode === 13){
-      var output = this.interpret(this.input);
+      var output = this.interpret(this.input, this.term);
       if(output == ""){
         this.term.reset();
       }
       else{
         this.term.write('\r\n');
-        this.term.writeln(output);
+        this.term.write(output);
       }
       this.input = '';
-      this.term.write('\r\n>');
+      this.term.write('\x1b[1;37m' + '\r\n>');
     }
     else if(event.keyCode === 8){
       this.term.write('\b \b');
@@ -53,11 +56,7 @@ export class CliComponent implements OnInit, AfterViewInit {
     this.input = '';
     this.term = new Terminal();
     this.term.open(this.terminalDiv.nativeElement);
-    this.term.write('\r\n>');
-    
-
-
-
+    this.term.write('\x1b[1;37m' + '\r\n>');
   }
 
   ngAfterViewInit(){
@@ -65,18 +64,20 @@ export class CliComponent implements OnInit, AfterViewInit {
 
 
     //this function taken in our input line, grabs the char that is a command, and runs it's corresponding code
-    interpret(line : string ){
+    interpret(line : string, currTerm: Terminal){
       var output : string = "";
-      switch(line[0]){
-        case "h": output = this.help(); break;
-        case "q": this.router.navigate(['']); break;
-        case "a": output = this.addClass(line); break;
-        case "e": output = this.editClass(line); break;
-        case "r": output = this.removeClass(line); break;
-        case "c": output = ""; break;
-        case "v": output = this.viewDiagram(); break;
-        case "x": output = "copy this: " + this.exportDiagram(); break;
-        default: output = "Error: Invalid Command. Type \"h\" for help"; break;
+      var input = line.replace(">","").split(" ");
+      switch(input[0]){
+        case "help": output = this.help(); break;
+        case "quit": this.router.navigate(['']); break;
+        case "add": output = this.addClass(line); break;
+        case "edit": output = this.editClass(line); break;
+        case "remove": output = this.removeClass(line); break;
+        case "clear": output = ""; break;
+        case "view": output = this.viewDiagram(); break;
+        case "export": output = "copy this: " + this.exportDiagram(); break;
+        case "clone": output = this.cloneClass(line); break;
+        default: output = '\x1b[1;31m' + "Error: Invalid Command. Type \"help\" for help"; break;
       }
       return output;
   }
@@ -90,13 +91,13 @@ export class CliComponent implements OnInit, AfterViewInit {
   //This prints the current diagram to the screen in a human-readable format
   //To "view" a diagram, you have to add a class through the gui first, as the CLI add is incomplete
   viewDiagram(){
-    var diagram:string = "\n";
+    var diagram:string = '\x1b[1;36m';
     let regex = /\n/gi
     for(var i = 0; i < this.service.allClasses.length; i++){
         diagram += "---------------------------------------------------\r\n";
-        diagram += "Class: " + this.service.allClasses[i].name + "\r\n";
-        diagram += "Methods: " + this.service.allClasses[i].methods.toString().replace(regex, " ") + "\r\n";
-        diagram += "Variables: " + this.service.allClasses[i].variables.toString().replace(regex, " ") + "\r\n";
+        diagram += "Class:\t\t" + this.service.allClasses[i].name + "\r\n";
+        diagram += "Methods:\t" + this.service.allClasses[i].methods.toString().replace(regex, " ") + "\r\n";
+        diagram += "Variables:\t" + this.service.allClasses[i].variables.toString().replace(regex, " ") + "\r\n";
         diagram += "---------------------------------------------------\r\n";
         diagram += "\n";
     }
@@ -114,24 +115,35 @@ export class CliComponent implements OnInit, AfterViewInit {
   addClass(line: string){
     var name = this.grabClassNameFromInput(line);
     if (name){
-      this.service.createNew(name, ["none"], ["none"]);
-      return "New Blank class \"" + name + "\" added! Type 'e' to edit attributes.";
+      var additon = this.selectClass(name);
+      if(additon){
+        return '\x1b[1;31m' + "Error: Class \"" + name + "\" already exists.";
+      }
+      else{
+        this.service.createNew(name, ["none"], ["none"]);
+        return '\x1b[1;32m' + "New Blank class \"" + name + "\" added! Type \'edit " + name + "\' to edit attributes.";
+      }
     }
     else{
-      return "Error: Invalid Syntax. \n\rThe format for this command is: a <class_name>";
+      return '\x1b[1;31m' + "Error:\tInvalid Syntax." + '\x1b[1;33m' + "\n\rFormat:\tadd <class_name>";
     }
   }
 
-  //this funtion removes a class wit the given name.
+  //this funtion removes a class with the given name.
   removeClass(line: string){
     var targetName = this.grabClassNameFromInput(line);
     if (targetName){
       var target = this.selectClass(targetName);
-      this.service.removeClassByIndex(this.service.allClasses.indexOf(target));
-      return "Class \"" + targetName + "\" has been deleted";
+      if(target){
+        this.service.removeClassByIndex(this.service.allClasses.indexOf(target));
+        return '\x1b[1;32m' + "Class \"" + targetName + "\" has been deleted";
+      }
+      else{
+        return '\x1b[1;31m' + "Errot: Class \"" + targetName + "\" not found. Type 'view' to view classes.";
+      }
     }
     else{
-      return "Error: Invalid Syntax. \n\rThe format for this command is: r <class_name>";
+      return '\x1b[1;31m' + "Error:\tInvalid Syntax." + '\x1b[1;33m' + "\n\rFormat:\tremove <class_name>";
     }
   }
 
@@ -140,15 +152,96 @@ export class CliComponent implements OnInit, AfterViewInit {
     return this.service.findClass(name);
   }
 
+  //this takes in a class name, and what you want
   editClass(line: string){
     var targetName = this.grabClassNameFromInput(line);
     var targetClass = this.selectClass(targetName);
-    var choppedLine = line.split("[");
-    var inputVariables = choppedLine[1];
-    inputVariables = inputVariables.replace("]","");
-    var inputMethods = choppedLine[2]; 
-    inputMethods = inputMethods.replace("]","");
-    return "input variables = " + inputVariables + "\n\r input  methods = " + inputMethods;
+    var addedClass = false;
+    if(!targetClass){
+      this.addClass(line);
+      var targetClass = this.selectClass(targetName);
+      addedClass = true;
+    }
+    var choppedLine = line.split(" ");
+    var bracketLine = line.split("[");
+    console.log("bl length: " + bracketLine.length);
+    if(choppedLine.length < 3 || bracketLine.length < 3){
+      return '\x1b[1;31m' + "Error:\tInvalid Syntax." + '\x1b[1;33m' + "\n\rFormat:\tedit <class_name> [var1,var2,...] [method1(),method2,...]";
+    }
+    else{
+      var inputVariables = bracketLine[1]
+      inputVariables = inputVariables.replace("]","");
+      this.updateAttribute(targetClass, inputVariables, "variables");
+
+      var inputMethods = bracketLine[2]; 
+      inputMethods = inputMethods.replace("]","");
+      this.updateAttribute(targetClass, inputMethods, "methods");
+
+      if(addedClass){
+        return '\x1b[1;33m' + "Class \"" + targetName + "\" did not exist, so it has been added.";
+      }
+      else{
+        return '\x1b[1;32m' + "Class \"" + targetName + "\" attributes have been updated.";
+      }
+    }
+
+  }
+
+  //this is a helper for edit. It updates the attribute of a given class with the passed in value.
+  updateAttribute(target: fullClass, value: string, field: string){
+    switch(field){
+      case "variables":
+        this.updateVariables(target, value);
+        break;
+      case "methods":
+        this.updateMethods(target, value);
+        break;
+      default:
+        break;
+    }
+  }
+
+  //this updates the variables of the passed in class with the passed in values
+  updateVariables(target: fullClass, values: string){
+    var valsArray = values.split(",");
+    target.variables.length = 0;
+    console.log("\'"+valsArray[0]+"\"");
+    if(valsArray[0] == " "){
+      target.variables.push("none");
+    }
+    else{
+      for(var i = 0; i < valsArray.length; i++){
+        target.variables.push(valsArray[i]);
+      }
+    }
+  }
+
+  //this updates the methods of the passed in class with the passed in methods
+  updateMethods(target: fullClass, methods: string){
+    var methodsArray = methods.split(",");
+    target.methods.length = 0;
+    if(methodsArray[0] == ""){
+      target.methods.push("none");
+    }
+    else{
+      for(var i = 0; i < methodsArray.length; i++){
+        target.methods.push(methodsArray[i]);
+      }
+    }
+  }
+
+  cloneClass(line: string){
+    var targetName = this.grabClassNameFromInput(line);
+    var targetClass = this.selectClass(targetName);
+    var choppedLine = line.split(" ");
+    console.log(choppedLine[2]);
+    if(choppedLine.length < 2){
+      return '\x1b[1;31m' + "Error:\tInvalid Syntax." + '\x1b[1;33m' + "\n\n\rFormat:\tclone <class_name> <clone_name>";
+    }
+    else{
+      this.service.createNew(choppedLine[2], targetClass.methods, targetClass.variables);
+      return '\x1b[1;32m' + "Class \"" + targetName + "\" cloned successfully.";
+    }
 
   }
 
@@ -156,13 +249,14 @@ export class CliComponent implements OnInit, AfterViewInit {
   help(){
     return `    Commands:\r
     ---------\r
-    a -> add a class\r
-    c -> clear the screen\r
-    e -> edit a class\r
-    h -> print help message\r
-    q -> quit and return to GUI\r
-    r -> remove a class\r
-    v -> view diagram\r
-    x -> export diagram\r`
+    add    -> add a class:\r
+    clear  -> clear the screen\r
+    clone  -> clone a class\r
+    edit   -> edit a class\r
+    export -> export diagram\r
+    help   -> print help message\r
+    quit   -> quit and return to GUI\r
+    remove -> remove a class\r
+    view   -> view diagram\r`
   }
 }
